chore(AppShell): migrate to rc6 and use ngc
1. Use ngc:
  - Generate metadata files.
  - Emit ESM modules in order in order to align with the core and provide easier tree-shaking.
2. Migrate to RC.6:
  - Migrate from deprecated APIs.
  - Update SystemJS config to use the Angular ES5 bundles.

Since the current build emits ESM there's a problem with the App Shell Runtime Parser which uses CommonJS internally in order to monkey-patch APIs of parse5 (which uses CommonJS as well).

As cleanest solution I'd suggest to move the Runtime Parse to a separate module once the SW plugin architecture is completed.