refactor(worker): rewrite of the worker driver logic
The previous SW relied on the worker update process (change bytes of worker script -> new worker with its own install/activate lifecycle) in order to update the application cached version. This is problematic for a few reasons:

* It requires changes to the worker script bytes any time the application is updated. This is difficult to achieve properly because it requires build integration to change the worker script. The previous worker used importScripts() under the assumption that changes in imported scripts would trigger the update process, but this has not yet become the case in the spec.
* It places control over updates in the browser's hands, thus limiting the ability to check for updates within the worker itself, or for the app to send push notifications when updates are available, or for the app to roll out updates only to selected users via the worker.
* It precludes ever having the worker serve multiple app versions at the same time.

Instead, a more flexible architecture is to have the worker maintain its cache of the application version itself. To make this possible, the worker needs to make several important decisions:

* Which version of the application is active?
* When a request comes in, should it serve a new version of the app or maintain the old?
* When should it check for updates?

This refactor refactors the Driver to route requests to a VersionWorker instance. The driver makes a decision for each request whether to reuse the existing VersionWorker or spin up a new one for a new version of the application. It checks for updates asynchronously each time the worker is started.