feat(worker): add Dynamic plugin for caching dynamic content
This commit adds plugins/dynamic for the SW, which supports caching of URLs that match
certain patterns. Different sets of URLs can be configured with different caching
strategies, via the `optimizeFor` mechanism. Two such strategies are implemented:

The Performance strategy optimizes for the fastest possible response by leveraging the
cache as much as possible. Requests are preferentially served from the cache, and optional
refresh-ahead behavior keeps the cache up to date when possible.

The Freshness strategy optimizes for minimal staleness, by always firing requests to the
network first. A configurable timeout allows for fallback on the cache if the network does
not respond quickly enough. Regardless of fallback, if the network request does complete,
the new value will be cached for future use.

Additionally, the service worker core has been refactored extensively in this change:

* Typescript 2.3 has been adopted, including a switch to included typings for UI thread SW
  APIs.
* Core APIs have been refactored to make writing the dynamic plugin code easier.