feat(worker): support coordinated updates
Ordinarily, the service worker does not activate a pending update until
all application tabs have been closed. For some applications this method
of background updates is acceptable. For others, particularly applications
with long-lived tabs, this can lead to staleness if the conditions for
a background update are never triggered.

This change introduces a channel by which the application can subscribe to
notifications of pending updates. The application can then prompt the user
about the update or choose to just go ahead and apply the update, via a new
method activateUpdate().