fix(worker): correctly emit falsy values sent by the ServiceWorker
Currently, there is check that completes the observable on falsy values
(`takeWhile(v => !!v)`). This was first introduced in [ffea4acfe][1], when there
was no other mechanism for completing the observable (and possibly all sent
values were objects). Then a different mechanism was introduced in
[563e7ac44][2], with a stricter check (`data === null`).

Since there are types of messages that send boolean values (e.g.
[`checkForUpdate`][3] or [`activateUpdate`][4]), it is necessary to not complete
the observable on `false`.

[1]: https://github.com/angular/mobile-toolkit/commit/ffea4acfe442da3af7050c37b0d7baf005efe834#diff-a223d9a67ca2c0a46e98a4f481f74944R133
[2]: https://github.com/angular/mobile-toolkit/commit/563e7ac4424e823bbde0867c7dbb527ea206d326#diff-a223d9a67ca2c0a46e98a4f481f74944R148
[3]: https://github.com/angular/mobile-toolkit/blob/f9d83026ea75af1f3a2848903e4072bce19cb5e9/service-worker/worker/src/worker/driver.ts#L850
[4]: https://github.com/angular/mobile-toolkit/blob/f9d83026ea75af1f3a2848903e4072bce19cb5e9/service-worker/worker/src/worker/driver.ts#L856